+++ protocols/gossipsub/src/behaviour/tests.rs
@@ -284,7 +284,7 @@ fn proto_to_message(rpc: &proto::RPC) -> Rpc {
     for message in rpc.publish.into_iter() {
         messages.push(RawMessage {
             source: message.from.map(|x| PeerId::from_bytes(&x).unwrap()),
-            data: message.data.unwrap_or_default(),
+            data: Bytes::from(message.data.unwrap_or_default()),
             sequence_number: message.seqno.map(|x| BigEndian::read_u64(&x)), // don't inform the application
             topic: TopicHash::from_raw(message.topic),
             signature: message.signature, // don't inform the application
@@ -1021,7 +1021,7 @@ fn test_handle_iwant_msg_cached() {
 
     let raw_message = RawMessage {
         source: Some(peers[11]),
-        data: vec![1, 2, 3, 4],
+        data: Bytes::from(vec![1, 2, 3, 4]),
         sequence_number: Some(1u64),
         topic: TopicHash::from_raw("topic"),
         signature: None,
@@ -1079,7 +1079,7 @@ fn test_handle_iwant_msg_cached_shifted() {
     for shift in 1..10 {
         let raw_message = RawMessage {
             source: Some(peers[11]),
-            data: vec![1, 2, 3, 4],
+            data: Bytes::from(vec![1, 2, 3, 4]),
             sequence_number: Some(shift),
             topic: TopicHash::from_raw("topic"),
             signature: None,
@@ -1552,7 +1552,7 @@ fn do_forward_messages_to_explicit_peers() {
 
     let message = RawMessage {
         source: Some(peers[1]),
-        data: vec![12],
+        data: Bytes::from(vec![12]),
         sequence_number: Some(0),
         topic: topic_hashes[0].clone(),
         signature: None,
@@ -1696,7 +1696,7 @@ fn no_gossip_gets_sent_to_explicit_peers() {
 
     let message = RawMessage {
         source: Some(peers[1]),
-        data: vec![],
+        data: Bytes::new(),
         sequence_number: Some(0),
         topic: topic_hashes[0].clone(),
         signature: None,
@@ -2166,7 +2166,7 @@ fn test_gossip_to_at_least_gossip_lazy_peers() {
     //receive message
     let raw_message = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![],
+        data: Bytes::new(),
         sequence_number: Some(0),
         topic: topic_hashes[0].clone(),
         signature: None,
@@ -2211,7 +2211,7 @@ fn test_gossip_to_at_most_gossip_factor_peers() {
     //receive message
     let raw_message = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![],
+        data: Bytes::new(),
         sequence_number: Some(0),
         topic: topic_hashes[0].clone(),
         signature: None,
@@ -2575,7 +2575,7 @@ fn test_do_not_gossip_to_peers_below_gossip_threshold() {
     // Receive message
     let raw_message = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![],
+        data: Bytes::new(),
         sequence_number: Some(0),
         topic: topics[0].clone(),
         signature: None,
@@ -2652,7 +2652,7 @@ fn test_iwant_msg_from_peer_below_gossip_threshold_gets_ignored() {
     // Receive message
     let raw_message = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![],
+        data: Bytes::new(),
         sequence_number: Some(0),
         topic: topics[0].clone(),
         signature: None,
@@ -2743,7 +2743,7 @@ fn test_ihave_msg_from_peer_below_gossip_threshold_gets_ignored() {
     //message that other peers have
     let raw_message = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![],
+        data: Bytes::new(),
         sequence_number: Some(0),
         topic: topics[0].clone(),
         signature: None,
@@ -2928,7 +2928,7 @@ fn test_ignore_rpc_from_peers_below_graylist_threshold() {
 
     let raw_message1 = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![1, 2, 3, 4],
+        data: Bytes::from(vec![1, 2, 3, 4]),
         sequence_number: Some(1u64),
         topic: topics[0].clone(),
         signature: None,
@@ -2938,7 +2938,7 @@ fn test_ignore_rpc_from_peers_below_graylist_threshold() {
 
     let raw_message2 = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![1, 2, 3, 4, 5],
+        data: Bytes::from(vec![1, 2, 3, 4, 5]),
         sequence_number: Some(2u64),
         topic: topics[0].clone(),
         signature: None,
@@ -2948,7 +2948,7 @@ fn test_ignore_rpc_from_peers_below_graylist_threshold() {
 
     let raw_message3 = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![1, 2, 3, 4, 5, 6],
+        data: Bytes::from(vec![1, 2, 3, 4, 5, 6]),
         sequence_number: Some(3u64),
         topic: topics[0].clone(),
         signature: None,
@@ -2958,7 +2958,7 @@ fn test_ignore_rpc_from_peers_below_graylist_threshold() {
 
     let raw_message4 = RawMessage {
         source: Some(PeerId::random()),
-        data: vec![1, 2, 3, 4, 5, 6, 7],
+        data: Bytes::from(vec![1, 2, 3, 4, 5, 6, 7]),
         sequence_number: Some(4u64),
         topic: topics[0].clone(),
         signature: None,